feat: seperates dual-range slider as different component
