+++ web-components/index.html
@@ -105,7 +105,23 @@
       </div>
 
       <div style="display: flex; margin: 2rem 1rem 1rem 1rem">
-        <folksonomy-properties />
+        <folksonomy-properties filter-placeholder="Filter properties...">
+          <span slot="title">Properties</span>
+          <span slot="description">
+            Open Food Facts allows anyone to reuse contributed properties or create new ones (see the
+            <a href="https://wiki.openfoodfacts.org/Folksonomy_Engine" target="_blank"
+              rel="noopener noreferrer">Folksonomy Engine project</a>). Here's the list of all contributed properties.
+          </span>
+          <span slot="loading-text">Loading properties...</span>
+          <span slot="error-text">Failed to load properties. Please try again later.</span>
+          <span slot="empty-state-text">No properties found.</span>
+          <span slot="download-button-text">Download CSV</span>
+          <span slot="reset-button-text">Clear filters</span>
+          <span slot="property-header-text">Property</span>
+          <span slot="documentation-header-text">Documentation</span>
+          <span slot="count-header-text">Count</span>
+          <span slot="values-header-text">Values</span>
+        </folksonomy-properties>
       </div>
 
       <div style="display: flex; margin: 2rem 1rem 1rem 1rem">
